requirements-local.txt
apache-airflow==2.6.3
apache-airflow-providers-amazon==7.3.0
apache-airflow-providers-google==10.4.0
black==23.1.0
ruff==0.0.280
